requirements.txt
# Core Data Science
pandas>=1.5.0
numpy>=1.21.0

# APIs
python-binance>=1.0.19
praw>=7.7.1

# Sentiment Analysis
textblob>=0.17.1
nltk>=3.8.1

# Dashboard & Visualization
streamlit>=1.28.0
plotly>=5.15.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.65.0
schedule>=1.2.0

# WebSocket for real-time data
websocket-client>=1.6.3